Update to fix recent changes
The <increment_version> value wasn't being used in export.ps1, and also the <exports_base_dir> hadn't been updated in the in the export_config.xml. I'll still need to add back the ability to override the increment_version with a switch though, after this jam.

Also VS Code decided to replace the spaces with tabs.